Dat: Fix data type of stars

+++ Parking.FindingSlotManagement.Application/Features/Manager/Parkings/ParkingManagement/Commands/CreateNewParking/CreateNewParkingCommandHandler.cs
@@ -66,7 +66,7 @@ namespace Parking.FindingSlotManagement.Application.Features.Manager.Parkings.Pa
                 parkingEntity.IsFull = false;
                 await _parkingRepository.Insert(parkingEntity);
                 parkingEntity.Code = "BX" + parkingEntity.ParkingId;
-                parkingEntity.Stars = 0;
+                parkingEntity.Stars = (float)0.0;
                 await _parkingRepository.Save();
                 var managerExist = await _userRepository.GetById(checkBusinessExist.UserId);
                 managerExist.ParkingId = parkingEntity.ParkingId;
